+++ capreolus/benchmark/robust04.py
@@ -19,3 +19,18 @@ class Robust04(Benchmark):
     topic_file = PACKAGE_PATH / "data" / "topics.robust04.301-450.601-700.txt"
     fold_file = PACKAGE_PATH / "data" / "rob04_cedr_folds.json"
     query_type = "title"
+
+
+@Benchmark.register
+class Robust04Yang19(Benchmark):
+    """Robust04 benchmark using the folds from Yang et al. [1]
+
+    [1] Wei Yang, Kuang Lu, Peilin Yang, and Jimmy Lin. 2019. Critically Examining the "Neural Hype": Weak Baselines and the Additivity of Effectiveness Gains from Neural Ranking Models. SIGIR 2019.
+    """
+
+    module_name = "robust04.yang19"
+    dependencies = [Dependency(key="collection", module="collection", name="robust04")]
+    qrel_file = PACKAGE_PATH / "data" / "qrels.robust2004.txt"
+    topic_file = PACKAGE_PATH / "data" / "topics.robust04.301-450.601-700.txt"
+    fold_file = PACKAGE_PATH / "data" / "rob04_yang19_folds.json"
+    query_type = "title"
