+++ capreolus/searcher/tests/test_searcher.py
@@ -1,8 +1,8 @@
 import os
 import numpy as np
 import pytest
-from profane import module_registry
 
+from capreolus import module_registry
 from capreolus.utils.trec import load_trec_topics
 from capreolus.benchmark import DummyBenchmark
 from capreolus.searcher import Searcher, BM25, BM25Grid
