+++ docs/quick.md
@@ -115,7 +115,7 @@ class TutorialTask(Task):
     config_spec = [ConfigOption("optimize", "map", "metric to maximize on the validation set")]
     dependencies = [
         Dependency(
-            key="benchmark", module="benchmark", name="robust04.yang19", provide_this=True, provide_children=["collection"]
+            key="benchmark", module="benchmark", name="nf", provide_this=True, provide_children=["collection"]
         ),
         Dependency(key="searcher1", module="searcher", name="BM25RM3"),
         Dependency(key="searcher2", module="searcher", name="SDM"),
